Started moving to boost.asio. We now have boost.system as library dependency. Added NetAddress as structure for ip/hostname and port. Transfered the LAN game finding code to boost and allowed IPv6.

+++ src/network/internet_gaming.h
@@ -170,7 +170,7 @@ private:
 	std::string pwd_;
 	bool reg_;
 	std::string meta_;
-	uint32_t port_;
+	uint16_t port_;
 
 	/// local clients name and rights
 	std::string clientname_;
